# Eliminate all allocations in parser