--- a/cabal.project
+++ b/cabal.project
@@ -3,5 +3,4 @@ packages: .
 source-repository-package
     type: git
     location: https://github.com/byteverse/bytesmith
-    tag: 2522d5b59e7bc684410176d9bea2309812ed7d40
-
+    tag: 37796446698354d0e310b4668c404437f34a1e29