--- a/cabal.project
+++ b/cabal.project
@@ -3,4 +3,4 @@ packages: .
 source-repository-package
     type: git
     location: https://github.com/byteverse/bytesmith
-    tag: 37796446698354d0e310b4668c404437f34a1e29
+    tag: d25a41304f83bc95dce89eebe800b4209ae8321a