Do not merge this. There is still some cleanup that needs to happen. This makes several changes:

* All fields in the `Url` data types are changed from `Word32` to `Int#`. This avoids some narrowing that was previously happening.
* The parser is moved into its own module (`Url.Rebind`) where nonstandard `>>=`, `>>`, and `pure` and imported from `Data.Bytes.Parser.Rebindable`. Unfortunately, `RebindableSyntax` does not actually work with a levity-polymorphic bind, so I desugared `do` manually.
* The data type `Url` is moved into `Url.Unsafe`. This had to be moved somewhere since `Url` imports `Url.Rebind`, and I didn't want to put it in `Url.Rebind`.
* Everything except for the `Url` data constructor is still exported from `Url`.

Manual inspection of the optimized core shows that the parser no longer allocates anything on the heap other than the resulting `Url`.

Before:

```
benchmarked url-bytes 1000
time                 227.0 μs   (224.9 μs .. 229.3 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 225.7 μs   (225.0 μs .. 226.9 μs)
std dev              2.910 μs   (2.147 μs .. 4.161 μs)
```

After:

```
benchmarked url-bytes 1000
time                 178.4 μs   (176.8 μs .. 180.5 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 178.1 μs   (177.4 μs .. 178.8 μs)
std dev              2.417 μs   (1.956 μs .. 2.969 μs)
```