@chessai 2020-02-23
Small nitpick, probably gets optimised away anyway, but instead of `I# x == I# y`, why not `isTrue# (x ==# y)`