@goolord 2020-02-25
alright, as far as I can tell this is 100% finished (sans the branch of bytesmith being merged), merging